+++ js/map.js
@@ -1,22 +1,7 @@
 import {
-  generatedAds
-} from './data.js';
-
-import {
-  generateCard
-} from './card.js';
-
-import {
   setAddressValue,
-  activateForm,
-  deactivateForm,
   resetButton
 } from './form.js';
-
-import {
-  activateMapFilters,
-  deactivateMapFilters
-} from './map-filters.js';
 
 const START_COORDINATES = {
   lat: 35.681700,
@@ -34,60 +19,52 @@ const AD_PIN = {
   iconAnchor: [20, 40],
 };
 
-const activatePage = () => {
-  activateMapFilters();
-  activateForm();
-};
+const addMap = (activatePage, deactivatePage) => {
+  deactivatePage();
+  const map = L.map('map-canvas')
+    .on('load', () => {
+      activatePage();
+    })
+    .setView(START_COORDINATES, START_ZOOM_LEVEL);
 
-const deactivatePage = () => {
-  deactivateMapFilters();
-  deactivateForm();
-};
+  L.tileLayer(
+    'https://{s}.tile.openstreetmap.org/{z}/{x}/{y}.png',
+    {
+      attribution: '&copy; <a href="https://www.openstreetmap.org/copyright">OpenStreetMap</a> contributors',
+    },
+  ).addTo(map);
 
-deactivatePage();
-const map = L.map('map-canvas')
-  .on('load', activatePage)
-  .setView(START_COORDINATES, START_ZOOM_LEVEL);
+  const mainPin = L.icon(MAIN_PIN);
 
-L.tileLayer(
-  'https://{s}.tile.openstreetmap.org/{z}/{x}/{y}.png',
-  {
-    attribution: '&copy; <a href="https://www.openstreetmap.org/copyright">OpenStreetMap</a> contributors',
-  },
-).addTo(map);
-
-const mainPin = L.icon(MAIN_PIN);
-
-const mainMarker = L.marker(
-  START_COORDINATES,
-  {
-    draggable: true,
-    icon: mainPin,
-  },
-);
-mainMarker.addTo(map);
-setAddressValue(START_COORDINATES);
-mainMarker.on('moveend', (evt) => {
-  const address = evt.target.getLatLng();
-  setAddressValue(address);
-});
-
-resetButton.addEventListener('click', () => {
+  const mainMarker = L.marker(
+    START_COORDINATES,
+    {
+      draggable: true,
+      icon: mainPin,
+    },
+  );
+  mainMarker.addTo(map);
   setAddressValue(START_COORDINATES);
-  mainMarker.setLatLng(START_COORDINATES);
-  map.setView(START_COORDINATES, START_ZOOM_LEVEL);
-});
+  mainMarker.on('moveend', (evt) => {
+    const address = evt.target.getLatLng();
+    setAddressValue(address);
+  });
 
-const adMarkerGroup = L.layerGroup().addTo(map);
+  resetButton.addEventListener('click', () => {
+    setAddressValue(START_COORDINATES);
+    mainMarker.setLatLng(START_COORDINATES);
+    map.setView(START_COORDINATES, START_ZOOM_LEVEL);
+  });
+  return map;
+};
 
-const createAdMarkers = (ad) => {
+const addMarkers = (map, location, card) => {
+  const adMarkerGroup = L.layerGroup().addTo(map);
+
   const adPin = L.icon(AD_PIN);
 
   const adMarker = L.marker(
-    {
-      lat: ad.location.lat,
-      lng: ad.location.lng,
-    },
+    location,
     {
       icon: adPin,
     },
@@ -96,13 +73,14 @@ const createAdMarkers = (ad) => {
   adMarker
     .addTo(adMarkerGroup)
     .bindPopup(
-      generateCard(ad),
+      card,
       {
         keepInView: true,
       },
     );
 };
 
-generatedAds.forEach((generatedAd) => {
-  createAdMarkers(generatedAd);
-});
+export {
+  addMap,
+  addMarkers
+};
